requirements.txt
flask
requests
google-cloud-vision
librosa
birdnetlib